In progress of testing PS4. Open loop locomotion
Progress from last commit: Fixed pin assignments to avoid reserved pins of ESP32.
Current issue:
1. Input shaping ramp function not working as expected. fluctuates between 0 and -1 when input to controller is 0
2. When writing to pins. Will get this error [gpio: gpio_set_level(227): GPIO output gpio_num error]
3. pwm value of 4094 being written to pin

+++ main/Globals.h
@@ -73,5 +73,5 @@ double previousWheelMotorInputs [4] = {0, 0, 0, 0};
 =                                PS4 GLOBAL VARIABLES                                    =
 ========================================================================================*/
 
-double PS4StickOutputs [4] = {0, 0, 0, 0};
+int PS4StickOutputs [4] = {0, 0, 0, 0};
 ControllerPtr myControllers[BP32_MAX_GAMEPADS];
